leo-7/LeoE-task7-task-202410111104/gen_test/AdvancedGameGridTest.java
package test;

import org.junit.Before;
import org.junit.Test;
import static org.junit.Assert.*;

public class AdvancedGameGridTest {

    @Test
    public void testPlayerInitialPosition() {
        char[][] grid = {
            {'.', '.', '.', '.', '.'},
            {'.', '.', 'P', '.', '.'},
            {'.', '.', '.', '.', '.'}
        };
        AdvancedGameGrid gameGrid = new AdvancedGameGrid(grid);
        assertEquals(1, gameGrid.getPlayerRow());
        assertEquals(2, gameGrid.getPlayerCol());
    }

    @Test(expected = IllegalArgumentException.class)
    public void testInvalidMoveDirection() {
        char[][] grid = {
            {'.', '.', '.', 'P', '.'},
            {'.', '.', '.', '.', '.'},
            {'.', '.', '.', '.', '.'}
        };
        AdvancedGameGrid gameGrid = new AdvancedGameGrid(grid);
        gameGrid.movePlayer('X');
    }

    @Test
    public void testMovePlayerValidMove() {
        char[][] grid = {
            {'.', '.', '.', 'P', '.'},
            {'.', '.', '.', '.', '.'},
            {'.', '.', '.', '.', '.'}
        };
        AdvancedGameGrid gameGrid = new AdvancedGameGrid(grid);
        gameGrid.movePlayer('S'); // Move down
        assertEquals(1, gameGrid.getPlayerRow());
        assertEquals(3, gameGrid.getPlayerCol());
    }

    @Test
    public void testMovePlayerInvalidMove() {
        char[][] grid = {
            {'P', 'E', '.', '.', '.'},
            {'.', '.', '.', '.', '.'},
            {'.', '.', '.', '.', '.'}
        };
        AdvancedGameGrid gameGrid = new AdvancedGameGrid(grid);
        gameGrid.movePlayer('D'); // Move right into enemy
        assertEquals(0, gameGrid.getPlayerRow());
        assertEquals(0, gameGrid.getPlayerCol());
    }

    @Test
    public void testEnemyMovement() {
        char[][] grid = {
            {'.', '.', '.', '.', '.'},
            {'.', 'E', '.', '.', '.'},
            {'.', '.', '.', '.', '.'}
        };
        AdvancedGameGrid gameGrid = new AdvancedGameGrid(grid);
        
        Random random = new Random(123); // To make enemy movement predictable
        gameGrid.mockRandom(random);
        
        gameGrid.moveEnemies();
        // Check enemy has moved correctly
        assertTrue(intermediaryPositionCheck(grid));
    }
    
    private boolean intermediaryPositionCheck(char[][] grid) {
        // Method to check correct intermediary position post-movement
        return (grid[0][1] == 'E' && grid[1][1] == '.') || 
               (grid[1][0] == 'E' && grid[1][1] == '.') ||
               (...); // Add additional expected positions     
    }

    @Test
    public void testDefeatEnemy() {
        char[][] grid = {
            {'.', '.', '.', 'P', '.'},
            {'.', 'E', '.', '.', '.'},
            {'.', '.', '.', '.', '.'}
        };
        AdvancedGameGrid gameGrid = new AdvancedGameGrid(grid);
        
        GameScoringSystem scoringSystem = gameGrid.getScoringSystem();
        
        gameGrid.movePlayer('S'); // Move down
        int expectedScore = 100; // Assuming defeating an enemy gives 100 points
        assertEquals(expectedScore, scoringSystem.getTotalScore());
    }

    @Test
    public void testGameLoop() {
        char[][] initialGrid = {
            {'.', '.', '.', 'P', '.', '.', '.', '.'},
            {'E', '.', '.', '.', 'E', '.', '.', '.'},
            {'.', '.', '.', 'E', '.', '.', '.', '.'}
        };
        AdvancedGameGrid gameGrid = new AdvancedGameGrid(initialGrid);
        for (int i = 0; i < 5; i++) {
            gameGrid.gameLoop();
        }
        // Verify state and scoring system post game loop execution
        ...
    }
}

// LevelLoaderTest.java
import org.junit.Test;
import static org.junit.Assert.*;
import java.io.File;
import java.io.IOException;
import java.io.PrintWriter;

